profiles-ext/i-sadav#6111 ajustes

+++ BusinessLogic/CatalogModule/Mapping/Catalogo_Cambio_Divisa.cs
@@ -11,8 +11,8 @@ namespace Business
         [PrimaryKey(Identity = true)]
         public int? Id_cambio { get; set; }
         public DateTime? Fecha { get; set; }
-        public double? Valor_de_compra { get; set; }
-        public double? Valor_de_venta { get; set; }
+        public decimal? Valor_de_compra { get; set; }
+        public decimal? Valor_de_venta { get; set; }
         public MonedaEnum? Moneda { get; set; }
         public Catalogo_Cambio_Divisa? GetDivisa(MonedaEnum? moneda = MonedaEnum.DOLAR)
         {
